requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
mongoengine==0.27.0
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
Werkzeug==2.3.7

# AI and Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
google-generativeai==0.3.2

# Location Services
geopy==2.4.1

# Data Visualization and Analytics
plotly==5.17.0

# Async Support
aiohttp==3.9.1

# Additional utilities
python-dateutil==2.8.20
scipy==1.11.2
python-dateutil==2.8.2
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
